# docs/openapi/generated/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@livetale.com
    name: LiveTale Team
  description: |
    LiveTale �꾨줈�앺듃�� API 紐낆꽭��
    �명꽣�숉떚釉� �ㅽ넗由ы뀛留곴낵 罹먮┃�� 梨꾪똿�� �꾪븳 醫낇빀 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LiveTale API
  version: 1.0.0
servers:
- description: Production server
  url: https://api.livetale.com/v1
- description: Staging server
  url: https://staging-api.livetale.com/v1
- description: Development server
  url: http://localhost:3000/v1
security:
- BearerAuth: []
- ApiKeyAuth: []
paths:
  /auth/signup:
    post:
      description: �덈줈�� �ъ슜�� 怨꾩젙�� �앹꽦�⑸땲��
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_auth_signup_post_request"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_auth_signup_post_201_response"
          description: �뚯썝媛��� �깃났
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "�섎せ�� �붿껌 (�대찓�� 以묐났, �좏슚�� 寃��� �ㅽ뙣 ��)"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �대찓�� �먮뒗 �ъ슜�먮챸 以묐났
      security: []
      summary: �뚯썝媛���
      tags:
      - Authentication
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Authentication
  /auth/login:
    post:
      description: �ъ슜�� �몄쬆�� �섑뻾�섍퀬 �≪꽭�� �좏겙�� 諛쒓툒�⑸땲��
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_auth_login_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_auth_login_post_200_response"
          description: 濡쒓렇�� �깃났
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆 �ㅽ뙣 (�섎せ�� �대찓�� �먮뒗 鍮꾨�踰덊샇)
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �덈Т 留롮� 濡쒓렇�� �쒕룄
      security: []
      summary: 濡쒓렇��
      tags:
      - Authentication
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Authentication
  /auth/logout:
    post:
      description: �ъ슜�� 濡쒓렇�꾩썐�� �섑뻾�섍퀬 �좏겙�� 臾댄슚�뷀빀�덈떎
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_auth_logout_post_200_response"
          description: 濡쒓렇�꾩썐 �깃났
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆�섏� �딆� �ъ슜��
      summary: 濡쒓렇�꾩썐
      tags:
      - Authentication
      x-accepts:
      - application/json
      x-tags:
      - tag: Authentication
...
components:
  schemas:
    User:
      example:
        subscription_tier: free
        is_active: true
        avatar_url: https://openapi-generator.tech
        updated_at: 2000-01-23T04:56:07.000+00:00
        nickname: nickname
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        email: email
        username: username
      properties:
        id:
          description: �ъ슜�� 怨좎쑀 ID
          format: uuid
          type: string
        email:
          description: �대찓�� 二쇱냼
          format: email
          type: string
        username:
          description: �ъ슜�먮챸
          maxLength: 20
          minLength: 3
          type: string
        nickname:
          description: �됰꽕��
          maxLength: 30
          type: string
        avatar_url:
          description: �꾨줈�� �대�吏� URL
          format: uri
          type: string
        created_at:
          description: 怨꾩젙 �앹꽦��
          format: date-time
          type: string
        updated_at:
          description: 留덉�留� �섏젙��
          format: date-time
          type: string
        is_active:
          description: 怨꾩젙 �쒖꽦�� �곹깭
          type: boolean
        subscription_tier:
          description: 援щ룆 �깃툒
          enum:
          - free
          - premium
          - pro
          type: string
      type: object
...
    Error:
      example:
        details: "{}"
        error: error
        message: message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      type: object
    _auth_signup_post_request:
      properties:
        email:
          description: �대찓�� 二쇱냼
          format: email
          type: string
        password:
          description: 鍮꾨�踰덊샇 (理쒖냼 8��)
          minLength: 8
          type: string
        username:
          description: �ъ슜�먮챸
          maxLength: 20
          minLength: 3
          type: string
        nickname:
          description: �됰꽕��
          maxLength: 30
          type: string
        avatar_url:
          description: �꾨줈�� �대�吏� URL (�좏깮�ы빆)
          format: uri
          type: string
      required:
      - email
      - nickname
      - password
      - username
      type: object
    _auth_signup_post_201_response:
      example:
        refresh_token: refresh_token
        user:
          subscription_tier: free
          is_active: true
          avatar_url: https://openapi-generator.tech
          updated_at: 2000-01-23T04:56:07.000+00:00
          nickname: nickname
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          email: email
          username: username
        token: token
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          description: JWT �≪꽭�� �좏겙
          type: string
        refresh_token:
          description: JWT 由ы봽�덉떆 �좏겙
          type: string
      type: object
    _auth_login_post_request:
      properties:
        email:
          description: �대찓�� 二쇱냼
          format: email
          type: string
        password:
          description: 鍮꾨�踰덊샇
          type: string
        remember_me:
          description: 濡쒓렇�� �곹깭 �좎� �щ�
          type: boolean
      required:
      - email
      - password
      type: object
    _auth_login_post_200_response:
      example:
        refresh_token: refresh_token
        user:
          subscription_tier: free
          is_active: true
          avatar_url: https://openapi-generator.tech
          updated_at: 2000-01-23T04:56:07.000+00:00
          nickname: nickname
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          email: email
          username: username
        expires_in: 0
        token: token
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          description: JWT �≪꽭�� �좏겙
          type: string
        refresh_token:
          description: JWT 由ы봽�덉떆 �좏겙
          type: string
        expires_in:
          description: �좏겙 留뚮즺 �쒓컙 (珥�)
          type: integer
      type: object
    _auth_logout_post_200_response:
      example:
        message: 濡쒓렇�꾩썐�섏뿀�듬땲��
      properties:
        message:
          example: 濡쒓렇�꾩썐�섏뿀�듬땲��
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey